Partial Refactor to implement POM with Serenity BDD - Test1 pass. HomePage SignInPage refactored. New package for the steps. Test2.java-Test6.java DON'T WORK - WIP.
