+++ src/test/java/PageObjects/CartPage.java
@@ -1,4 +1,4 @@
-package Pages;
+package PageObjects;
 
 import net.serenitybdd.core.pages.PageObject;
 import org.openqa.selenium.By;
