+++ src/test/java/PageObjects/LogInPage.java
@@ -1,4 +1,4 @@
-package Pages;
+package PageObjects;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
