+++ src/test/java/steps/Test2.java
@@ -1,59 +1,60 @@
-package starter;
+package steps;
 
-import Pages.CartPage;
-import Pages.HomePage;
+import PageObjects.*;
 import net.serenitybdd.junit.runners.SerenityRunner;
 import net.thucydides.core.annotations.Managed;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
+import org.junit.FixMethodOrder;
 import org.junit.runner.RunWith;
+import org.junit.runners.MethodSorters;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.support.ui.WebDriverWait;
 
-import java.time.Duration;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
+import java.time.Duration;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 @RunWith(SerenityRunner.class)
-public class Test5 {
+@FixMethodOrder(MethodSorters.NAME_ASCENDING)
+public class Test2 {
+    private static final String TEST_NAME="start121zs";
+    private static final String TEST_PWD="marte45szs";
 
     @Managed
     WebDriver driver;
     WebDriverWait wait;
-    HomePage homePage;
-    CartPage cartPage;
+    private HomePage homePage;
+    private LogInPage logInPage;
 
     @Before
     public void setUp() {
-        homePage = new HomePage(driver);
-        cartPage = new CartPage(driver);
         wait = new WebDriverWait(driver, Duration.ofSeconds(3));
+        homePage = new HomePage(driver);
+        logInPage = new LogInPage(driver);
     }
 
     @Test
-    public void test5_remove_s6_from_cart() {
+    public void test2_login() {
         homePage.navigateTo();
-        homePage.clickOnPhones();
-        homePage.clickOnPhoneS6();
-        homePage.addProductToCart();
-        cartPage.navigateTo();
-        assertThat(cartPage.getProductNamesFromCart()).contains("Samsung galaxy s6");
-        cartPage.removeProductFromCart("Samsung galaxy s6");
-        assertThat(cartPage.getProductNamesFromCart().size()).isLessThan(1);
+        homePage.clickOnLogIn();
+        logInPage.logInUser(TEST_NAME, TEST_PWD);
+        String loggedUserName = homePage.getWelcomeMessage();
+        assertThat(loggedUserName).isEqualToIgnoringCase("Welcome " + TEST_NAME);
     }
 
     @After
-    public void tearDown(){
+    public void close_browser() {
         String osName = System.getProperty("os.name");
         String osVersion = System.getProperty("os.version");
         LocalDate today = LocalDate.now();
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("EEEE dd',' MMMM yyyy");
         String formattedDate = today.format(formatter);
 
-        System.out.println("\nReport name : TEST5");
+        System.out.println("\nReport name : TEST2");
         System.out.println("O.S name    : " + osName + " version " + osVersion);
         System.out.println("Date        : " + formattedDate + "\n");
         driver.quit();
