+++ src/test/java/steps/Test3.java
@@ -1,7 +1,7 @@
-package starter;
+package steps;
 
-import Pages.CartPage;
-import Pages.HomePage;
+//import io.cucumber.junit.CucumberSerenityRunner; TODO: USED WITH CUCUMBER
+import PageObjects.HomePage;
 import net.serenitybdd.junit.runners.SerenityRunner;
 import net.thucydides.core.annotations.Managed;
 import org.junit.After;
@@ -15,40 +15,49 @@ import java.time.Duration;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 
+import static org.assertj.core.api.Assertions.*;
 @RunWith(SerenityRunner.class)
-public class Test6 {
+public class Test3 {
 
     @Managed
     WebDriver driver;
     WebDriverWait wait;
     private HomePage homePage;
-    private CartPage cartPage;
+
     @Before
-    public void setUp(){
+    public void setUp() {
         homePage = new HomePage(driver);
-        cartPage = new CartPage(driver);
         wait = new WebDriverWait(driver, Duration.ofSeconds(3));
     }
 
     @Test
-    public void test6_do_checkout(){
+    public void test3_navigate_between_different_products() {
         homePage.navigateTo();
+        homePage.clickOnPhones();
+        assertThat(homePage.hasPhones()).isTrue();
+        assertThat(homePage.hasLaptops()).isFalse();
+        assertThat(homePage.hasMonitors()).isFalse();
+
         homePage.clickOnLaptops();
-        homePage.clickOnLaptopVaioI7();
-        homePage.addProductToCart();
-        cartPage.navigateTo();
-        cartPage.buyItemsInCart();
+        assertThat(homePage.hasPhones()).isFalse();
+        assertThat(homePage.hasLaptops()).isTrue();
+        assertThat(homePage.hasMonitors()).isFalse();
+
+        homePage.clickOnMonitors();
+        assertThat(homePage.hasPhones()).isFalse();
+        assertThat(homePage.hasLaptops()).isFalse();
+        assertThat(homePage.hasMonitors()).isTrue();
     }
 
     @After
-    public void tearDown(){
+    public void tearDown() {
         String osName = System.getProperty("os.name");
         String osVersion = System.getProperty("os.version");
         LocalDate today = LocalDate.now();
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("EEEE dd',' MMMM yyyy");
         String formattedDate = today.format(formatter);
 
-        System.out.println("\nReport name : TEST6");
+        System.out.println("\nReport name : TEST3");
         System.out.println("O.S name    : " + osName + " version " + osVersion);
         System.out.println("Date        : " + formattedDate + "\n");
         driver.quit();
