+++ src/test/resources/serenity.conf
@@ -1,14 +1,14 @@
 drivers {
     windows {
-    webdriver.chrome.driver = src/test/resources/drivers/windows/chromedriver
+        webdriver.chrome.driver = src/test/resources/drivers/windows/chromedriver
     }
     linux {
-    webdriver.chrome.driver = src/test/resources/drivers/linux/chromedriver
+        webdriver.chrome.driver = src/test/resources/drivers/linux/chromedriver
     }
     mac64 {
         webdriver.chrome.driver = src/test/resources/drivers/mac64/chromedriver
-        }
+    }
     mac_arm64 {
-            webdriver.chrome.driver = src/test/resources/drivers/mac_arm64/chromedriver
-            }
+        webdriver.chrome.driver = src/test/resources/drivers/mac_arm64/chromedriver
+    }
 }
